+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* logic for setting layer type to table
+
 ## [2.9.1] - 12-07-2017
 ### Fixed
 * Position of capabilities